reset completed epochs when switching to/from interactive mode
